refactor: Customer의 닉네임을 VO로 변경
